Move some of the definitions in LegacyNNDefinitions.cpp closer to sites
All of these definitions are no longer "legacy" as their CPU
implementations have been ported to ATen.  There are probably some
layers of indirection that could be reduced here, but for now just do a
minor but unlikely to break things cleanup.

The last thing in LegacyNNDefinitions truly is still in THCUNN and can't
be removed.

Signed-off-by: Edward Z. Yang <ezyang@fb.com>

ghstack-source-id: 5b10a12aef9f436930e35deaf915ccad0ce78725
Pull Request resolved: https://github.com/pytorch/pytorch/pull/37531