reenable quantized test_compare_tensor_scalar
Pull Request resolved: https://github.com/pytorch/pytorch/pull/37422

The test was failing because in fbcode the version of hypothesis was too old to know
about the width parameter, and it was trying to generate values larger than float32.  The fix
is to explicitly set the defaults of the floats range for old versions of hypothesis.

For now, reenable the test and see what breaks in CI

Differential Revision: [D21310336](https://our.internmc.facebook.com/intern/diff/D21310336/)
ghstack-source-id: 103500358