[TensorExpr] Fix the shape info check in the TE fuser pass.
Previously we checked if a node's inputs and outputs have shape
info only when we tried to merge this node into an existing fusion
group, but we didn't check it for the first node in the group. This PR
fixes that. It was causing a failure on test_milstm_cuda, which is now
fixed.

ghstack-source-id: e347306e040c5f33dc1ea8987ca46c56efc654fe
Pull Request resolved: https://github.com/pytorch/pytorch/pull/37882