[TensorPipe] Use the new multi-payload message API
In D21209901 TensorPipe added support for a vector of payloads inside each message, instead of a single one, so that users with multiple payloads can send them separately as they are instead of having to copy them into a new block of contiguous memory. The PyTorch agent is using the old API, which is preventing us from deleting it. This change has no effects on over-the-wire format and thus on performance.

Pull Request resolved: https://github.com/pytorch/pytorch/pull/37919


ghstack-source-id: 103572164

Differential Revision: [D21425536](https://our.internmc.facebook.com/intern/diff/D21425536/)