[TensorExpr] TensorExprKernel: don't do any compilation or lowering in run().
The input JIT graph has all the information we need to perform the
entire compilation at the construction time. We don't need to postpone
any steps until the execution time. Also, from the graph we always know
what device we will be executing on and thus we don't need to have a
CodeGen cache in TensorExprKernel - we always have one and only one
CodeGen.

ghstack-source-id: e1fc41cd2bc7320b3cd90dc311084d7d206b306f
Pull Request resolved: https://github.com/pytorch/pytorch/pull/37948